Improving documentation. Still a lot to add.
Everything that we add to Groundhog that changes an existing component
or adds a new functionality should be documented from now on.

Javadocs aren't enough and examples are very helpful.

So, for example, if a feature of fetching Pull Requests is added, there
should be at least one example about its usage in the README or in the
project's Wiki https://github.com/spgroup/groundhog/wiki